Updated code.
updated question 4. Will put the automator from question 4 auto on at
the end.

Here is the code. It works out most of the issues, still needs to work
out number of events that occurred in 2011 but that should be easy.
Only bad part is that it loads sloooow. Going to work on that more
tomorrow morning, because a million rows will take about 1500 minutes.
Too long!